Sample app updates
