+++ WebViewApp-ObjC/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,7 +1,6 @@
 {
   "images" : [
     {
-      "filename" : "ObjCSample.jpg",
       "idiom" : "universal",
       "platform" : "ios",
       "size" : "1024x1024"
