+++ WebViewApp-Swift/WebViewApp-Swift/Assets.xcassets/AppIcon.appiconset/Contents.json
