+++ WebViewApp-Swift/WebViewApp-Swift/Extensions/StringExtensions.swift
@@ -1,6 +1,6 @@
 //
 //  StringExtensions.swift
-//  SampleApplication
+//  WebViewApp-Swift
 //
 //  Created by Todd Bohman on 8/2/22.
 //
