+++ WebViewApp-Swift/WebViewApp-Swift/HostController.swift
@@ -1,6 +1,6 @@
 //
 //  HostController.swift
-//  TunnelMAMTestApp2
+//  WebViewApp-Swift
 //
 //  Created by Todd Bohman on 8/9/22.
 //
