+++ WebViewApp-Swift/WebViewApp-Swift/Modifiers/SafeColor.swift
@@ -1,6 +1,6 @@
 //
 //  SafeColor.swift
-//  TunnelMAMTestApp2
+//  WebViewApp-Swift
 //
 //  Created by Todd Bohman on 8/29/22.
 //
