+++ WebViewApp-Swift/WebViewApp-Swift/ViewModels/WebViewModel.swift
@@ -1,6 +1,6 @@
 //
 //  WebViewModel.swift
-//  SampleApplication
+//  WebViewApp-Swift
 //
 //  Created by Todd Bohman on 8/1/22.
 //
