+++ WebViewApp-Swift/WebViewApp-Swift/WebView/WebViewContainer.swift
@@ -1,6 +1,6 @@
 //
 //  WebViewContainer.swift
-//  SampleApplication
+//  WebViewApp-Swift
 //
 //  Created by Todd Bohman on 8/1/22.
 //
